+++ app/page.tsx
@@ -2,6 +2,7 @@ import { auth } from "@/auth";
 import { SignInButton, JoinGuildButton } from "@/components/auth-buttons";
 import Playground from "@/components/playground";
 import UserInfo from "@/components/user-info";
+import { getCeremonyTimeString } from "@/lib/ceremony-data";
 import { getLatestOverallPassportId } from "@/lib/get-latest-passport";
 import { getOptimizedLatestPassportImage } from "@/lib/get-optimized-latest-passport-image";
 import { MySession, OptimizedLatestPassportImage } from "@/types/types";
@@ -55,23 +56,36 @@ export default async function Home({
 						/>
 						Passport Data Pages
 					</h1>
-					<div className="rounded-sm border-[3px] border-amber-400 flex flex-col justify-center w-full md:w-10/12 gap-2 p-3 sm:p-4 my-4 mx-auto break-inside-avoid shadow-amber-600 shadow-blocks-sm font-main">
-						<p className="text-base">
-							Use this website to create your passport data page. 🛂
-						</p>
-						<p className="text-base">
-							The info can be whatever you want. Some people have put their real
-							info, others have put totally wacky stuff. Half of the passports
-							have a cat set as their portrait.
-						</p>{" "}
-						{userId ? (
-							<p className="text-base">
-								When you&#39;re ready to register for a passport ceremony, check
-								the checkbox to be assigned a passport number. Your passport
-								will be printed & ready to assemble at the next ceremony.
+					{!latestPassport ? (
+						<>
+							<div className="rounded-sm border-[3px] border-amber-400 flex flex-col justify-center w-full md:w-10/12 gap-2 p-3 sm:p-4 my-4 mx-auto break-inside-avoid shadow-amber-600 shadow-blocks-sm font-main">
+								<p className="text-base">
+									Use this website to create your passport data page. 🛂
+								</p>
+								<p className="text-base">
+									The info can be whatever you want. Some people have put their real
+									info, others have put totally wacky stuff. Half of the passports
+									have a cat set as their portrait.
+								</p>{" "}
+								{userId ? (
+									<p className="text-base">
+										When you&#39;re ready to register for a passport ceremony, check
+										the checkbox to be assigned a passport number. Your passport
+										will be printed & ready to assemble at the next ceremony.
+									</p>
+								) : null}
+							</div>
+						</>
+					) : (
+						<div className="rounded-sm border-[3px] border-amber-400 flex flex-col justify-center w-full md:w-10/12 gap-2 p-3 sm:p-4 my-4 mx-auto break-inside-avoid shadow-amber-600 shadow-blocks-sm font-main">
+							<p>
+								SOMETHING GOES HERE (MAKE IT NOT AMBER TOO) Ceremony Date:{" "}<br />
+								{getCeremonyTimeString(latestPassport.ceremony_time)}<br />
+								This will show for all &quot;existing&quot; passports, so it
+								will be here when returning as well
 							</p>
-						) : null}
-					</div>
+						</div>
+					)}
 					{!session?.user?.email ? (
 						<div className="rounded-sm border-[3px] border-red-400 flex flex-col justify-center w-full md:w-10/12 gap-4 p-3 sm:p-4 my-4 mx-auto break-inside-avoid shadow-red-600 shadow-blocks-sm font-main">
 							<p>
